requirements.txt
aiolimiter==1.1.0
anyio==4.0.0
APScheduler==3.10.4
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
cryptography==41.0.5
exceptiongroup==1.1.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.1
hyperframe==6.0.1
idna==3.4
lxml==4.9.3
pycparser==2.21
python-docx==1.1.0
python-telegram-bot==20.6
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
tornado==6.3.3
typing_extensions==4.8.0
tzlocal==5.2
